Finalized Application with CSS applied.
The application is finalized using a reset css file and a a styles css file. There is a little error: The buttons don't keep always in the
bottom of the card, it means if the paragraph get a line minus, the
button of the current section will get up from the others.